+++ ui/components/navigation/main-menu.mjs
@@ -27,6 +27,7 @@ import {
   OpenLockIcon,
   PackageIcon,
   PlusCircleIcon,
+  WindowIcon,
   QuestionIcon,
   RightIcon,
   SearchIcon,
@@ -48,7 +49,6 @@ import { Link } from 'components/link'
  * Shared props for icons in the sidebar/navigation
  */
 export const iconProps = { className: 'w-6 h-6 shrink-0 grow-0', stroke: 1.25 }
-
 
 /*
  * Object to map icons to page
@@ -90,6 +90,7 @@ const icons = {
   networks: WifiIcon,
   nodes: ServersIcon,
   notes: TipIcon,
+  oss: WindowIcon,
   pkgs: PackageIcon,
   presets: CheckCircleIcon,
   proxy: Traefik,
@@ -119,35 +120,35 @@ export const links = {
     r: 'user',
     audit: {
       t: 'Audit',
-      r: 'user'
+      r: 'user',
     },
     events: {
       t: 'Events',
-      r: 'user'
+      r: 'user',
     },
     checks: {
       t: 'Health Checks',
-      r: 'user'
+      r: 'user',
     },
     logs: {
       t: 'Logs',
-      r: 'user'
+      r: 'user',
     },
     metrics: {
       t: 'Metrics',
-      r: 'user'
+      r: 'user',
     },
     notes: {
       t: 'Notes',
-      r: 'user'
+      r: 'user',
     },
     search: {
       t: 'Search Dashboards',
-      r: 'user'
+      r: 'user',
     },
     custom: {
       t: 'Custom Dashboards',
-      r: 'user'
+      r: 'user',
     },
   },
   inventory: {
@@ -155,15 +156,19 @@ export const links = {
     r: 'operator',
     hosts: {
       t: 'Hosts',
-      r: 'user'
+      r: 'user',
     },
     ips: {
       t: 'IP Addresses',
-      r: 'user'
+      r: 'user',
     },
     macs: {
       t: 'MAC Addresses',
-      r: 'user'
+      r: 'user',
+    },
+    oss: {
+      t: 'Operating Systems',
+      r: 'user',
     },
   },
   settings: {