+++ ui/pages/boards/metrics/[host]/index.mjs
@@ -1,19 +1,19 @@
 import { PageWrapper } from 'components/layout/page-wrapper.mjs'
 import { ContentWrapper } from 'components/layout/content-wrapper.mjs'
-import { LogsIcon } from 'components/icons.mjs'
-import { HostLogsTable } from 'components/boards/logs.mjs'
+import { StatusIcon } from 'components/icons.mjs'
+import { HostMetricsTable } from 'components/boards/metrics.mjs'
 
-export default function DashboardsHostLogsPage({ host }) {
+export default function DashboardsHostMetricsPage({ host }) {
   const meta = {
-    title: 'Cached host logs',
-    page: ['boards', 'logs', host],
-    Icon: LogsIcon,
+    title: 'Cached host metrics',
+    page: ['boards', 'metrics', host],
+    Icon: StatusIcon,
   }
 
   return (
     <PageWrapper {...meta}>
       <ContentWrapper {...meta}>
-        <HostLogsTable host={host} />
+        <HostMetricsTable host={host} />
       </ContentWrapper>
     </PageWrapper>
   )